refactor: rename 'storage' to 'base'
This avoids importing from 'alliance_platform.storage.storage' which
is unpleasant
